+++ clabe/types/clabes_legacy/errors.py
@@ -1,11 +1,4 @@
-from .helpers import is_pydantic_v1_installed
-
-if is_pydantic_v1_installed():
-    from pydantic.errors import PydanticValueError
-else:
-    from pydantic_core import PydanticCustomError
-    class PydanticValueError(PydanticCustomError):
-        ...
+from pydantic.errors import PydanticValueError
 
 
 class BankCodeValidationError(PydanticValueError):
